Changed handling of NULL
