+++ exporter.go
@@ -2,12 +2,7 @@ package trdsql
 
 import (
 	"context"
-	"encoding/hex"
-	"fmt"
 	"log"
-	"strconv"
-	"time"
-	"unicode/utf8"
 )
 
 // Exporter is the interface for processing query results.
@@ -89,29 +84,4 @@ func (e *WriteFormat) ExportContext(ctx context.Context, db *DB, query string) e
 	}
 
 	return e.Writer.PostWrite()
-}
-
-// ValString converts database value to string.
-func ValString(v interface{}) string {
-	switch t := v.(type) {
-	case nil:
-		return ExportNULL
-	case string:
-		return t
-	case []byte:
-		if ok := utf8.Valid(t); ok {
-			return string(t)
-		}
-		return `\x` + hex.EncodeToString(t)
-	case int:
-		return strconv.Itoa(t)
-	case int32:
-		return strconv.FormatInt(int64(t), 10)
-	case int64:
-		return strconv.FormatInt(t, 10)
-	case time.Time:
-		return t.Format(time.RFC3339)
-	default:
-		return fmt.Sprint(v)
-	}
 }
